Moved encryption rotation methods into the rotation module to clean us the class
