+++ lib/decryption.rb
@@ -31,20 +31,4 @@ class Decryption
     third_reversal.join("").to_s
     fourth_reversal.join("").to_s
   end
-
-  # def first_reversal
-  #   generate_reversal(@a_key, 0)
-  # end
-  #
-  # def second_reversal
-  #   generate_reversal(@b_key, 1)
-  # end
-  #
-  # def third_reversal
-  #   generate_reversal(@c_key, 2)
-  # end
-  #
-  # def fourth_reversal
-  #   generate_reversal(@d_key, 3)
-  # end
 end
