+++ lib/encryption.rb
@@ -1,8 +1,10 @@
 require 'date'
 require_relative 'generator'
+require_relative 'rotation'
 
 class Encryption
 include Generator
+include Rotation
 
   attr_reader :a_key,
               :b_key,
@@ -28,21 +30,5 @@ include Generator
     second_rotation.join("").to_s
     third_rotation.join("").to_s
     fourth_rotation.join("").to_s
-  end
-
-  def first_rotation
-    generate_rotation(@a_key, 0)
-  end
-
-  def second_rotation
-    generate_rotation(@b_key, 1)
-  end
-
-  def third_rotation
-    generate_rotation(@c_key, 2)
-  end
-
-  def fourth_rotation
-    generate_rotation(@d_key, 3)
   end
 end
